Updates 3
I have done a test on the Submit text on the button in the RegistrationForm.js and it passed, but there is a failed test in the App that I haven't created any test for it, it says something to do with the styled-components. I also have some basic styling.